Implement orders and payments service. Improve raffles service.

+++ server/services/orders-service/src/main/java/com/raffleease/orders_service/Kafka/Messages/Failure/FailureNotification.java
@@ -6,7 +6,7 @@ import jakarta.validation.constraints.NotNull;
 
 public record FailureNotification(
         @NotNull(message = "Must indicate order")
-        Integer orderId,
+        Long orderId,
 
         @NotNull(message = "Must indicate new order status")
         OrderStatus status
