+++ server/services/orders-service/src/main/java/com/raffleease/orders_service/Kafka/Messages/Success/SuccessNotification.java
@@ -1,6 +1,6 @@
 package com.raffleease.orders_service.Kafka.Messages.Success;
 
-import com.raffleease.orders_service.Customers.DTO.CustomerDTO;
+import com.raffleease.orders_service.Customers.CustomerDTO;
 import com.raffleease.orders_service.Payments.DTO.PaymentData;
 import jakarta.validation.constraints.NotNull;
 import lombok.Builder;
@@ -9,7 +9,7 @@ import lombok.Builder;
 public record SuccessNotification(
 
         @NotNull(message = "Must provide an order id")
-        Integer orderId,
+        Long orderId,
 
         @NotNull(message = "Must provide payment data")
         PaymentData paymentData,
